pyproject.toml
[tool.poetry]
name = "firstserver"
version = "0.1.0"
description = "Servidor criado com flask para estudos py."
authors = ["Johnnick F. Landim <johnnick.landim@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
httpx = "^0.23.1"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
mysql-connector = "^2.2.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pyclean = "^2.2.0"
blue = "^0.9.1"
prospector = "^1.7.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
